#48 - Added -measure-with-source operation and test
