+++ src/test/java/org/opencds/cqf/TestBase.java
@@ -32,6 +32,7 @@ public class TestBase {
         measureTests.populationMeasureBCS();
         measureTests.populationMeasureCCS();
         measureTests.populationMeasureCOL();
+        measureTests.bundleSourceDataMeasure_COL();
     }
 
     @Test
